feat: common navbar
